refactor: colombia data access in separate package
